Move session.Manager interface to caller side
