+++ session/create_consumer_test.go
@@ -20,12 +20,13 @@ package session
 import (
 	"testing"
 
+	"github.com/mysteriumnetwork/node/identity"
 	"github.com/stretchr/testify/assert"
 )
 
 var consumer = createConsumer{
 	CurrentProposalID: 101,
-	SessionManager:    &ManagerFake{},
+	SessionManager:    &managerFake{},
 }
 
 func TestConsumer_UnknownProposal(t *testing.T) {
@@ -61,4 +62,20 @@ func TestConsumer_Success(t *testing.T) {
 		},
 		sessionResponse,
 	)
+}
+
+// managerFake represents fake manager usually useful in tests
+type managerFake struct{}
+
+var fakeConfig = struct {
+	Param1 string
+	Param2 int
+}{
+	"string-param",
+	123,
+}
+
+// Create function creates and returns fake session
+func (manager *managerFake) Create(peerID identity.Identity) (Session, error) {
+	return Session{"new-id", fakeConfig, peerID}, nil
 }
